Moved the four impls of `BitDenotation` to their own module, `mod impls`.
